requirements.txt
langchain-core
langchain-google-genai
python-dotenv
langchain-community
langchain_openai
typing-extensions
fastapi
pydantic
requests
bs4
pathlib
langgraph
uvicorn
Jinja2
google-cloud-secret-manager
google-cloud-resource-manager
google-cloud-build
google-auth
google-cloud-logging
google-cloud-iam
google-cloud-storage
PyJWT[crypto]
pymongo